api/requirements.txt
fastapi
sqlalchemy
PyJWT
passlib
python-multipart
uvicorn
pg8000
bcrypt
